common, mount: Fix iolimits issue in Windows
The iolimits option is a feature to configure hard limits on the
throughput of a client instance. Old Windows Client version was not able
to perform well altogether with this feature, it got stuck every time.

Digging deep in the code, specifically to the TokenBucket::attempt
function, it was noticed that when the throughput limit and the interval
time between calls to verifyClockSteadiness function was very low, the
whole iolimits issue was triggered. Even more specifically, the result
from std::min(cost, budget_) was a floating point value less than 1 and
the final use of this response is an integer, which means that it got
truncated to 0, but the TokenBucket::budget_ variable was still reducing
its value by the floating point previous result, thus reducing to
probably 0.

Final solution is to change the return type of TokenBucket::attempt and
the result variable to uint64_t. This will prevent the previous
truncating issue, allowing TokenBucket::budget_ add up results less than
1 without resetting its value to 0.

This commit also contains the related changes in the unittests.

+++ src/common/token_bucket.h
@@ -45,7 +45,7 @@ public:
 	double budgetCeil() const;
 
 	// Try to satisfy the request. Return 'cost' or less, but not less than 0. Require cost > 0
-	double attempt(SteadyTimePoint now, double cost);
+	uint64_t attempt(SteadyTimePoint now, double cost);
 
 private:
 	void updateBudget(SteadyTimePoint now);
